rename datadogSites to datacenters
